+++ src/components/gitHub/gitHubSection.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const GitHubSection = () => {
+  return <div>GitHubSection</div>;
+};
