+++ src/components/gitHub/index.ts
@@ -1,0 +1,1 @@
+export { GitHubSection } from "./gitHubSection";
