+++ src/pages/index.tsx
@@ -10,8 +10,8 @@ import { ErrorWrapper } from "src/components/atom/error";
 import { Hero } from "src/components/atom/hero";
 import { Title } from "src/components/atom/title";
 import { BlogCards } from "src/components/card/blogCards";
-import { GitHubReps } from "src/components/card/gitHubCards";
 import { PortfolioCardSlider } from "src/components/card/portfolioCardSlider";
+import { GitHubReps } from "src/components/gitHub/gitHubCards";
 import { TwitterSec } from "src/components/twitter";
 import { Layout } from "src/layout";
 import { useViewportSize } from "src/lib/mantine";
